+++ crates/swc_html_minifier/tests/fixture.rs
@@ -73,7 +73,7 @@ fn test_minify_document(input: PathBuf) {
             gen.emit(&document).unwrap();
         }
 
-        NormalizedOutput::from(html_str)
+        NormalizedOutput::new_raw(html_str)
             .compare_to_file(&output)
             .unwrap();
 
@@ -179,7 +179,7 @@ fn test_minify_document_fragment(input: PathBuf) {
             gen.emit(&document_fragment).unwrap();
         }
 
-        NormalizedOutput::from(html_str)
+        NormalizedOutput::new_raw(html_str)
             .compare_to_file(&output)
             .unwrap();
 
@@ -237,7 +237,7 @@ fn test_minify_recovery(input: PathBuf) {
             gen.emit(&document).unwrap();
         }
 
-        NormalizedOutput::from(html_str)
+        NormalizedOutput::new_raw(html_str)
             .compare_to_file(&output)
             .unwrap();